% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{report_dat}
\alias{report_dat}
\title{report_dat}
\format{
A data frame with five variables: \code{safetyreportid}, \code{age},
\code{year}, \code{D} and \code{E}.
}
\usage{
report_dat
}
\description{
Catetgorization of reports for a drug and adverse event
within the FDA's adverse event reporting system for pediatric patients
defined within the NICHD defined developmental stages
}
\details{
the drug indicator is for reporting of Methylphenidate and
the event indicator is for the reporting of Paranoia
}
\keyword{datasets}
